Xtrain_Diet_Meal/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
billiard==4.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.58.2
django-auth==0.1.9
django-cors-headers==4.3.0
django-pictures==1.1.3
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
executing==2.0.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
ifaddr==0.2.0
ipython==8.16.1
jedi==0.19.1
kombu==5.3.2
matplotlib-inline==0.1.6
multidict==6.0.4
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
pictures==0.7
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
py-eureka-client==0.11.8
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
replicate==0.15.5
requests==2.31.0
requests-oauthlib==1.3.1
self==2020.12.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
